requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.42
asyncpg==0.30.0
python-dotenv==1.1.1
pydantic==2.11.7
